[Refactor] #7 - 구조체 모델로 이동, 미사용 import문 삭제
