+++ BookSearchProject_BY/BookSearchProject_BY/Scenes/SavedBooks/SavedBooksView.swift
@@ -83,6 +83,8 @@ final class SavedBooksView: UIView {
     }
     
     private func setupButtonAction() {
+        let searchButtonClicked = searchBar.rx.searchButtonClicked
+            .withLatestFrom(searchBar.rx.text.orEmpty)
         deleteAllButton.addTarget(self, action: #selector(deleteAllButtonTapped), for: .touchUpInside)
         addBookButton.addTarget(self, action: #selector(addBookButtonTapped), for: .touchUpInside)
     }
