+++ BookSearchProject_BY/BookSearchProject_BY/Scenes/Search/SearchView.swift
@@ -153,19 +153,6 @@ final class SearchView: UIView {
     }
 }
 
-enum SearchSection: Int, CaseIterable {
-    case recentBooks
-    case searchResults
-    
-    var title: String {
-        switch self {
-        case .recentBooks: return "최근 본 책"
-        case .searchResults: return "검색 결과"
-        }
-    }
-}
-
-
 extension SearchView: UICollectionViewDataSource {
     func numberOfSections(in collectionView: UICollectionView) -> Int {
         return SearchSection.allCases.count
